Enum change, removed '_'
git-svn-id: https://svn.imm.dtu.dk/se2/svn/e14-groupC@181 a3cd46ed-ea8e-4e4b-8fda-82757e0212fd

+++ Software/Prototype1/PrototypeApp/src/com/example/prototypeapp/WebServiceConnection.java
@@ -13,7 +13,7 @@ public class WebServiceConnection {
 	//private static final String BASE_URL = "http://se-se2-e14-glassfish-c.compute.dtu.dk:8080/TechnologyExperiment/rest";
 	//private static final String BASE_URL = "http://10.16.163.41:8080/TechnologyExperiment/rest";
 	
-	public static Event invokeAddEventWebServer(String mac, float value, EventType eventType) {
+	public static Event invokeAddEventWebServer(String mac, int value, EventType eventType) {
 		final String url = BASE_URL + "/events/addEventByMac?mac={mac}&value={value}&eventType={eventType}";
 		final Map<String,String> parameters = new HashMap<String,String>();
 		parameters.put("mac", String.valueOf(mac));
